pyproject.toml
[project]
name = "cop-coffee-assistant"
version = "0.1.0"
description = "AI Coffee Shop Assistant with LangGraph, Groq, and Telegram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.0.10",
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "langchain-community>=0.0.10",
    "python-telegram-bot>=20.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
]